+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 <!--next-version-placeholder-->
 
+## v0.3.0 (2025-03-17)
+
+### Feature
+
+* **reset:** Add `--pop` option to download a backup before resetting ([`969fdba`](https://github.com/educationwarehouse/edwh-devdb-plugin/commit/969fdba6b2d905791fcdacd912fffa8bf440ad9f))
+
+## v0.2.0 (2025-03-15)
+
+### Feature
+
+* Allow `--compress` in `devdb.snapshot`, determine amount of threads automatically (cores - 1) instead of hard-coded (3) ([`5817f1c`](https://github.com/educationwarehouse/edwh-devdb-plugin/commit/5817f1c048b6b6c0f5cf1be69c94d5240f8d1d3a))
+
 ## v0.1.1 (2025-03-07)
 
 ### Documentation